Extract shared TranscriptSegments speaker parser from DriveSource
